# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FITVO API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/auth/register/professional": {
      "post": {
        "summary": "Cadastra um profissional (cria conta + tenant SOLO)",
        "tags": [
          "auth"
        ],
        "description": "Cria conta + tenant SOLO + perfil profissional (D-045) e dispara a verificacao de e-mail.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "E-mail de login (único) — D-042."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "allOf": [
                      {
                        "type": "string",
                        "pattern": "[A-Za-z]"
                      },
                      {
                        "type": "string",
                        "pattern": "\\d"
                      }
                    ],
                    "description": "Senha em claro — mín. 8, ao menos 1 letra e 1 número."
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Nome civil — deriva tenant.name e uso fiscal/documento."
                  },
                  "socialName": {
                    "description": "Nome social (exibido no lugar do civil quando preenchido) — spec §3.1.",
                    "type": "string",
                    "minLength": 1
                  },
                  "gender": {
                    "description": "Gênero/identidade — opcional (spec §3.1).",
                    "type": "string",
                    "enum": [
                      "MULHER_CIS",
                      "HOMEM_CIS",
                      "MULHER_TRANS",
                      "HOMEM_TRANS",
                      "NAO_BINARIO",
                      "OUTRO",
                      "PREFIRO_NAO_INFORMAR"
                    ]
                  },
                  "document": {
                    "type": "string",
                    "pattern": "^\\d+$",
                    "description": "CPF ou CNPJ, só dígitos (D-043)."
                  },
                  "documentType": {
                    "type": "string",
                    "enum": [
                      "CPF",
                      "CNPJ"
                    ]
                  },
                  "whatsapp": {
                    "type": "string",
                    "pattern": "^\\d{11}$",
                    "description": "WhatsApp — 11 dígitos (DDD + celular), só números (máscara é UI)."
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
                    "description": "Data de nascimento (YYYY-MM-DD) — maioridade obrigatória (D-044)."
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "cep": {
                        "type": "string",
                        "pattern": "^\\d{8}$",
                        "description": "CEP — 8 dígitos, só números (máscara é UI)."
                      },
                      "logradouro": {
                        "type": "string",
                        "minLength": 1
                      },
                      "numero": {
                        "type": "string",
                        "minLength": 1
                      },
                      "complemento": {
                        "type": "string"
                      },
                      "bairro": {
                        "type": "string",
                        "minLength": 1
                      },
                      "cidade": {
                        "type": "string",
                        "minLength": 1
                      },
                      "state": {
                        "type": "string",
                        "enum": [
                          "AC",
                          "AL",
                          "AM",
                          "AP",
                          "BA",
                          "CE",
                          "DF",
                          "ES",
                          "GO",
                          "MA",
                          "MG",
                          "MS",
                          "MT",
                          "PA",
                          "PB",
                          "PE",
                          "PI",
                          "PR",
                          "RJ",
                          "RN",
                          "RO",
                          "RR",
                          "RS",
                          "SC",
                          "SE",
                          "SP",
                          "TO"
                        ],
                        "description": "UF do endereço."
                      },
                      "country": {
                        "default": "BR",
                        "description": "País do endereço (default BR).",
                        "type": "string"
                      }
                    },
                    "required": [
                      "cep",
                      "logradouro",
                      "numero",
                      "bairro",
                      "cidade",
                      "state"
                    ],
                    "description": "Endereço da pessoa (D-044)."
                  },
                  "specialtyId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Especialidade reivindicada no signup (D-137)."
                  },
                  "councilDocument": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 20,
                    "pattern": "^[A-Za-z0-9/-]+$",
                    "description": "Registro no conselho (CREF/CRN/CRM) — validado so em formato (D-138)."
                  },
                  "councilState": {
                    "type": "string",
                    "enum": [
                      "AC",
                      "AL",
                      "AM",
                      "AP",
                      "BA",
                      "CE",
                      "DF",
                      "ES",
                      "GO",
                      "MA",
                      "MG",
                      "MS",
                      "MT",
                      "PA",
                      "PB",
                      "PE",
                      "PI",
                      "PR",
                      "RJ",
                      "RN",
                      "RO",
                      "RR",
                      "RS",
                      "SC",
                      "SE",
                      "SP",
                      "TO"
                    ],
                    "description": "UF do conselho profissional (D-126)."
                  },
                  "acceptedTerms": {
                    "type": "object",
                    "properties": {
                      "termsOfUse": {
                        "type": "boolean",
                        "enum": [
                          true
                        ],
                        "description": "Aceite explícito dos Termos de Uso."
                      },
                      "privacyPolicy": {
                        "type": "boolean",
                        "enum": [
                          true
                        ],
                        "description": "Aceite explícito da Política de Privacidade."
                      }
                    },
                    "required": [
                      "termsOfUse",
                      "privacyPolicy"
                    ],
                    "description": "Aceite dos termos no cadastro (D-025) — ambos obrigatoriamente `true`."
                  }
                },
                "required": [
                  "email",
                  "password",
                  "name",
                  "document",
                  "documentType",
                  "whatsapp",
                  "birthDate",
                  "address",
                  "specialtyId",
                  "councilDocument",
                  "councilState",
                  "acceptedTerms"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string",
                          "description": "Nome civil — uso fiscal/documento; NÃO exibir quando há nome social."
                        },
                        "displayName": {
                          "type": "string",
                          "description": "Nome de exibição (socialName ?? name) — spec §3.1."
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "displayName"
                      ],
                      "additionalProperties": false
                    },
                    "tokens": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "accessExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "refreshExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "accessExpiresAt",
                        "refreshExpiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "account",
                    "tokens"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/login": {
      "post": {
        "summary": "Autentica por e-mail e senha (rate limited)",
        "tags": [
          "auth"
        ],
        "description": "Emite access + refresh (D-029). Limitado a 5 tentativas por minuto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "E-mail de login (único) — D-042."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string",
                          "description": "Nome civil — uso fiscal/documento; NÃO exibir quando há nome social."
                        },
                        "displayName": {
                          "type": "string",
                          "description": "Nome de exibição (socialName ?? name) — spec §3.1."
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "displayName"
                      ],
                      "additionalProperties": false
                    },
                    "tokens": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "accessExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "refreshExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "accessExpiresAt",
                        "refreshExpiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "account",
                    "tokens"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "Rotaciona o refresh token",
        "tags": [
          "auth"
        ],
        "description": "Rotacao a cada uso com deteccao de reuso (D-029).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokens": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        },
                        "accessExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "refreshExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "accessToken",
                        "refreshToken",
                        "accessExpiresAt",
                        "refreshExpiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "tokens"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Encerra a sessao atual (revoga o refresh)",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "summary": "Conta autenticada (a partir do access token)",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string",
                      "description": "Nome civil — uso fiscal/documento; NÃO exibir quando há nome social."
                    },
                    "displayName": {
                      "type": "string",
                      "description": "Nome de exibição (socialName ?? name) — spec §3.1."
                    },
                    "emailVerified": {
                      "type": "boolean"
                    },
                    "profileComplete": {
                      "type": "boolean",
                      "description": "Perfil completo (spec §5) — derivado no servidor; nunca recalcular no cliente."
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "displayName",
                    "emailVerified",
                    "profileComplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/reset-password": {
      "post": {
        "summary": "Redefine a senha e revoga as sessoes",
        "tags": [
          "auth"
        ],
        "description": "Consome o token de uso unico e revoga todas as sessoes da conta (D-029).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Token de uso único recebido por e-mail."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Senha em claro (mín. 8)."
                  }
                },
                "required": [
                  "token",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}